added a base class for db_clients

+++ arc_scaper/_tests/test_arc_rest_model.py
@@ -1,19 +1,19 @@
 from arc_rest_model import ArcRestModel
 import requests
-from db_clients.mssql_db_client import SqlServerClient
 from os import path
-from _tests.mock_helper import mock_get_return, db_conn, get_generated_sql_expected, save_test_to_file
+from _tests.mock_helper import mock_get_return, get_db_conn, get_generated_sql_expected, save_test_to_file
 import pytest
 
 
 @pytest.fixture(autouse=True)
-def arc_rest_model():
+def arc_rest_model(db_type):
     uri = "http://geoportal.menlhk.go.id/arcgis/rest"
     database = 'test_Indonesia_menlhk'
-    sql_conn = SqlServerClient(db_conn)
-    return ArcRestModel(uri, database, sql_conn)
+    sql_client = get_db_conn(db_type)
+    return ArcRestModel(uri, database, sql_client)
 
 
+@pytest.mark.parametrize("db_type", ['mssql', 'mysql', 'postgresql'])
 def test_arc_rest_model_init(monkeypatch, arc_rest_model):
     monkeypatch.setattr(requests, 'get', mock_get_return)
     arm = arc_rest_model
@@ -23,14 +23,14 @@ def test_arc_rest_model_init(monkeypatch, arc_rest_model):
     assert len(arm.errors) == 2
     assert len(arm.arc_services) == 9
 
-
+@pytest.mark.parametrize("db_type", ['mssql', 'mysql', 'postgresql'])
 def test_arc_rest_model_generate_sql(monkeypatch, arc_rest_model):
     monkeypatch.setattr(requests, 'get', mock_get_return)
     service = arc_rest_model
     sql = service.generate_sql()
 
-    save_test_to_file("arc_service_generate_sql.txt", sql)
-    sql_expected = get_generated_sql_expected("arc_service_generate_sql.txt")
+    save_test_to_file(f"{arc_rest_model.db_client.name}_arc_service_generate_sql.sql", sql)
+    sql_expected = get_generated_sql_expected(f"{arc_rest_model.db_client.name}_arc_service_generate_sql.sql")
 
     assert sql.replace("\n", '').replace("\r", "").replace("\t", '').strip() \
            == sql_expected.replace("\n", '').replace("\r", "").replace("\t", '').strip()
@@ -42,13 +42,14 @@ def test_arc_rest_model_generate_sql(monkeypatch, arc_rest_model):
         assert False
 
 
+@pytest.mark.parametrize("db_type", ['mssql', 'mysql', 'postgresql'])
 def test_arc_rest_model_generate_sql_extra(monkeypatch, arc_rest_model):
     monkeypatch.setattr(requests, 'get', mock_get_return)
     service = arc_rest_model
     sql = service.generate_sql_extra()
 
-    save_test_to_file("arc_service_generate_sql_extra.txt", sql)
-    sql_expected = get_generated_sql_expected("arc_service_generate_sql_extra.txt")
+    save_test_to_file(f"{arc_rest_model.db_client.name}_arc_service_generate_sql_extra.sql", sql)
+    sql_expected = get_generated_sql_expected(f"{arc_rest_model.db_client.name}_arc_service_generate_sql_extra.sql")
 
     assert sql.replace("\n", '').replace("\r", "").replace("\t", '').strip() \
            == sql_expected.replace("\n", '').replace("\r", "").replace("\t", '').strip()
@@ -60,13 +61,14 @@ def test_arc_rest_model_generate_sql_extra(monkeypatch, arc_rest_model):
         assert False
 
 
+@pytest.mark.parametrize("db_type", ['mssql', 'mysql', 'postgresql'])
 def test_arc_rest_model_generate_sql_preamble(monkeypatch, arc_rest_model):
     monkeypatch.setattr(requests, 'get', mock_get_return)
     service = arc_rest_model
     sql = service.generate_sql_preamble()
 
-    save_test_to_file("arc_service_generate_sql_preamble.txt", sql)
-    sql_expected = get_generated_sql_expected("arc_service_generate_sql_preamble.txt")
+    save_test_to_file(f"{arc_rest_model.db_client.name}_arc_service_generate_sql_preamble.sql", sql)
+    sql_expected = get_generated_sql_expected(f"{arc_rest_model.db_client.name}_arc_service_generate_sql_preamble.sql")
 
     assert sql.strip() == sql_expected.strip()
 
@@ -77,28 +79,36 @@ def test_arc_rest_model_generate_sql_preamble(monkeypatch, arc_rest_model):
         assert False
 
 
-def test_save_sql_script(monkeypatch, arc_rest_model):
-    file = "C:\\Users\\npongo\Dropbox\\PythonProjects\\ArcMapService\\arc_scaper\\_tests\\test_output\\Script.sql"
+@pytest.mark.parametrize("db_type, file", [('mssql', "C:\\Users\\npongo\\Dropbox\\PythonProjects\\ArcMapService\\arc_scaper\\_tests\\test_output\\MSSql_Script.sql"),
+                                           ('mysql', "C:\\Users\\npongo\\Dropbox\\PythonProjects\\ArcMapService\\arc_scaper\\_tests\\test_output\\MySql_Script.sql"),
+                                           ('postgresql', "C:\\Users\\npongo\\Dropbox\\PythonProjects\\ArcMapService\\arc_scaper\\_tests\\test_output\\PostgreSql_Script.sql")])
+def test_save_sql_script(monkeypatch, arc_rest_model, file):
     monkeypatch.setattr(requests, 'get', mock_get_return)
     arm = arc_rest_model
-
     arm.save_sql_script(file)
 
     assert path.isfile(file)
 
 
-def test_run_sql_script(monkeypatch, arc_rest_model):
+@pytest.mark.parametrize("db_type, db_master", [('mssql', 'master'),
+                                                ('mysql', 'sys'),
+                                                ('postgresql', 'postgres')])
+def test_run_sql_script(monkeypatch, db_type, db_master, arc_rest_model):
     monkeypatch.setattr(requests, 'get', mock_get_return)
     arm = arc_rest_model
 
     db = arc_rest_model.db_client.db_conn.get('database', 'Test_Indonesia_menlhk')
-    sql = f"IF DB_ID('{db}') IS NOT NULL DROP DATABASE {db}"
-    arc_rest_model.db_client.db_conn['database'] = 'master'
+    sql = f'drop database {db}'
+    if db_type == "mysql":
+        sql = f"DROP DATABASE IF EXISTS {db}"
+    if db_type == 'mssql':
+        sql = f"IF DB_ID('{db}') IS NOT NULL DROP DATABASE {db}"
+    arc_rest_model.db_client.db_conn['database'] = db_master
     arc_rest_model.db_client.exec_non_query(sql, autocommit=True)
     arc_rest_model.db_client.db_conn['database'] = db
 
     try:
-        arm.run_sql()
+        arm.run_sql(db_master=db_master)
     except:
         assert False
